refact: move class file to appropriate package
