+++ src/main/kotlin/com/dobby/backend/application/service/OauthService.kt
@@ -2,8 +2,8 @@ package com.dobby.backend.application.service
 
 import com.dobby.backend.application.usecase.FetchGoogleUserInfoUseCase
 import com.dobby.backend.application.usecase.FetchNaverUserInfoUseCase
-import com.dobby.backend.presentation.api.dto.request.GoogleOauthLoginRequest
-import com.dobby.backend.presentation.api.dto.request.NaverOauthLoginRequest
+import com.dobby.backend.presentation.api.dto.request.auth.GoogleOauthLoginRequest
+import com.dobby.backend.presentation.api.dto.request.auth.NaverOauthLoginRequest
 import com.dobby.backend.presentation.api.dto.response.auth.OauthLoginResponse
 import org.springframework.stereotype.Service
