+++ src/main/kotlin/com/dobby/backend/application/service/SignupService.kt
@@ -1,7 +1,7 @@
 package com.dobby.backend.application.service
 
 import com.dobby.backend.application.usecase.ParticipantSignupUseCase
-import com.dobby.backend.presentation.api.dto.request.ParticipantSignupRequest
+import com.dobby.backend.presentation.api.dto.request.signup.ParticipantSignupRequest
 import com.dobby.backend.presentation.api.dto.response.signup.SignupResponse
 import jakarta.transaction.Transactional
 import org.springframework.stereotype.Service
