+++ src/main/kotlin/com/dobby/backend/application/usecase/FetchGoogleUserInfoUseCase.kt
@@ -6,11 +6,11 @@ import com.dobby.backend.infrastructure.config.properties.GoogleAuthProperties
 import com.dobby.backend.infrastructure.database.entity.enum.MemberStatus
 import com.dobby.backend.infrastructure.database.entity.enum.ProviderType
 import com.dobby.backend.infrastructure.database.repository.MemberRepository
-import com.dobby.backend.infrastructure.feign.GoogleAuthFeignClient
-import com.dobby.backend.infrastructure.feign.GoogleUserInfoFeginClient
+import com.dobby.backend.infrastructure.feign.google.GoogleAuthFeignClient
+import com.dobby.backend.infrastructure.feign.google.GoogleUserInfoFeginClient
 import com.dobby.backend.infrastructure.token.JwtTokenProvider
-import com.dobby.backend.presentation.api.dto.request.GoogleOauthLoginRequest
-import com.dobby.backend.presentation.api.dto.request.GoogleTokenRequest
+import com.dobby.backend.presentation.api.dto.request.auth.GoogleOauthLoginRequest
+import com.dobby.backend.presentation.api.dto.request.auth.GoogleTokenRequest
 import com.dobby.backend.presentation.api.dto.response.auth.GoogleTokenResponse
 import com.dobby.backend.presentation.api.dto.response.auth.OauthLoginResponse
 import com.dobby.backend.util.AuthenticationUtils
