+++ src/main/kotlin/com/dobby/backend/application/usecase/FetchNaverUserInfoUseCase.kt
@@ -6,11 +6,11 @@ import com.dobby.backend.infrastructure.config.properties.NaverAuthProperties
 import com.dobby.backend.infrastructure.database.entity.enum.MemberStatus
 import com.dobby.backend.infrastructure.database.entity.enum.ProviderType
 import com.dobby.backend.infrastructure.database.repository.MemberRepository
-import com.dobby.backend.infrastructure.feign.NaverAuthFeignClient
-import com.dobby.backend.infrastructure.feign.NaverUserInfoFeignClient
+import com.dobby.backend.infrastructure.feign.naver.NaverAuthFeignClient
+import com.dobby.backend.infrastructure.feign.naver.NaverUserInfoFeignClient
 import com.dobby.backend.infrastructure.token.JwtTokenProvider
-import com.dobby.backend.presentation.api.dto.request.NaverOauthLoginRequest
-import com.dobby.backend.presentation.api.dto.request.NaverTokenRequest
+import com.dobby.backend.presentation.api.dto.request.auth.NaverOauthLoginRequest
+import com.dobby.backend.presentation.api.dto.request.auth.NaverTokenRequest
 import com.dobby.backend.presentation.api.dto.response.auth.NaverTokenResponse
 import com.dobby.backend.presentation.api.dto.response.auth.OauthLoginResponse
 import com.dobby.backend.util.AuthenticationUtils
