+++ src/main/kotlin/com/dobby/backend/infrastructure/feign/google/GoogleUserInfoFeginClient.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend.infrastructure.feign
+package com.dobby.backend.infrastructure.feign.google
 
 import com.dobby.backend.presentation.api.dto.response.auth.GoogleInfoResponse
 import org.springframework.cloud.openfeign.FeignClient
