+++ src/main/kotlin/com/dobby/backend/infrastructure/feign/naver/NaverAuthFeignClient.kt
@@ -1,6 +1,6 @@
-package com.dobby.backend.infrastructure.feign
+package com.dobby.backend.infrastructure.feign.naver
 
-import com.dobby.backend.presentation.api.dto.request.NaverTokenRequest
+import com.dobby.backend.presentation.api.dto.request.auth.NaverTokenRequest
 import com.dobby.backend.presentation.api.dto.response.auth.NaverTokenResponse
 import org.springframework.cloud.openfeign.FeignClient
 import org.springframework.web.bind.annotation.PostMapping
