+++ src/main/kotlin/com/dobby/backend/presentation/api/controller/SignupController/ParticipantSignupController.kt
@@ -1,9 +1,8 @@
 package com.dobby.backend.presentation.api.controller.SignupController
 
 import com.dobby.backend.application.service.SignupService
-import com.dobby.backend.infrastructure.database.entity.enum.RoleType
 import com.dobby.backend.infrastructure.database.entity.enum.RoleType.*
-import com.dobby.backend.presentation.api.dto.request.ParticipantSignupRequest
+import com.dobby.backend.presentation.api.dto.request.signup.ParticipantSignupRequest
 import com.dobby.backend.presentation.api.dto.response.signup.SignupResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
