+++ src/main/kotlin/com/dobby/backend/presentation/api/dto/request/auth/MemberRefreshTokenRequest.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend.presentation.api.dto.request
+package com.dobby.backend.presentation.api.dto.request.auth
 
 import io.swagger.v3.oas.annotations.media.Schema
 import jakarta.validation.constraints.NotBlank
