+++ src/main/kotlin/com/dobby/backend/presentation/api/dto/request/auth/NaverTokenRequest.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend.presentation.api.dto.request
+package com.dobby.backend.presentation.api.dto.request.auth
 
 import com.fasterxml.jackson.annotation.JsonProperty
